Fix clang++ (version 20.1.8) compile errors.
error: a template argument list is expected after a name prefixed by the
template keyword [-Wmissing-template-arg-list-after-template-kw]

error: ‘strncmp’ was not declared in this scope

Personally I'd use `std::strncmp`, but for backwards compatibility
with C, strncmp is also made available in global namespace. Not even
sure if that is guaranteed by C++ though (probably not).

Also, I removed the `#include <string>` because this file doesn't
use `std::string` anywhere. It DOES use `string`, but whatever that is,
it wouldn't compile anyway with or without <string> if that wasn't
already defined elsewhere (again, one should use `std::string` if it
is the C++ standard string).