Use quoted include for tests/Common.h.
The directory tests/ has a single header (Common.h) that
is included only by .cpp files in the same directory.

Using the quote form makes the preprocessor first look for
the header in the directory of the .cpp file that includes it,
and only then the include paths. In general one should use
the quoted form when including a header from the same directory,
or a subdirectory (it avoids name collisions with externally
installed (system) headers).

In this case we can remove the include directory of the root
of the project now, too, because tests/Common.h is the only
header that required that.