Improve build.sh.
Adds options:
  -j, --parallel=N   build using N jobs in parallel.
  -h, --help         print a short overview of options.

Do not create the `build` directory (cmake already does that).
Do not cd into the `build` directory (now you can run build.sh
from any directory using a full path to it; although the build
directory will be created in the current directory from where
you run it).

Use cmake to build the project and to install it. This is more
portable (for example supports a portable --parallel N option).